refactor: move adapter out as module
