+++ backend/solontax-g1-management-adapter/src/main/java/solontax/g1/management/SolontaxG1ManagementApplication.java
@@ -1,0 +1,13 @@
+package solontax.g1.management;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "solontax.g1.management")
+public class SolontaxG1ManagementApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SolontaxG1ManagementApplication.class, args);
+    }
+
+}
