+++ backend/solontax-g1-management-adapter/src/main/resources/application.properties
@@ -1,0 +1,14 @@
+spring.application.name=solontax-g1-management-adapter
+server.port=${SPRING_SERVER_PORT:8080}
+# Database Configuration
+spring.datasource.url=${SPRING_DATASOURCE_URL:jdbc:postgresql://localhost:5430/solontax_g1}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME:postgres}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:postgres}
+spring.datasource.driver-class-name=org.postgresql.Driver
+# JPA Configuration
+spring.jpa.hibernate.ddl-auto=${SPRING_JPA_DDL_AUTO:update}
+spring.jpa.show-sql=${SPRING_JPA_SHOW_SQL:true}
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.format_sql=true
+# Timezone Configuration
+spring.jpa.properties.hibernate.jdbc.time_zone=UTC
