+++ backend/solontax-g1-management-adapter/src/test/java/solontax/g1/management/core/adapter/SolontaxG1ManagementAdapterApplicationTests.java
@@ -1,0 +1,15 @@
+package solontax.g1.management.core.adapter;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class SolontaxG1ManagementAdapterApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
