+++ backend/solontax-g1-management-adapter/src/test/resources/application-test.properties
