+++ backend/solontax-g1-management-core/src/main/java/solontax/g1/management/core/application/dto/PersonDto.java
@@ -1,0 +1,24 @@
+package solontax.g1.management.core.application.dto;
+
+import lombok.*;
+
+import java.util.UUID;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class PersonDto {
+    private UUID id;
+
+    private String firstName;
+
+    private String lastName;
+
+    private Long age;
+
+    private Long taxNumber;
+
+    private Long taxDebt;
+}
