+++ backend/solontax-g1-management-core/src/main/java/solontax/g1/management/core/domain/model/Person.java
@@ -1,0 +1,25 @@
+package solontax.g1.management.core.domain.model;
+
+import lombok.*;
+
+import java.time.LocalDate;
+import java.util.UUID;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Person {
+    private UUID id;
+
+    private String firstName;
+
+    private String lastName;
+
+    private LocalDate dateOfBirth;
+
+    private Long taxNumber;
+
+    private Long taxDebt;
+}
