--- backend/solontax-g1-management/src/main/java/solontax/g1/management/exception/CommonExceptionDto.java
@@ -1,19 +1,0 @@
-package solontax.g1.management.exception;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@Setter
-public class CommonExceptionDto {
-    private HttpStatus status;
-    private String message;
-}
