--- backend/solontax-g1-management/src/main/resources/application.properties
@@ -1,31 +1,0 @@
-spring.application.name=solontax-g1-management
-server.port=${SPRING_SERVER_PORT:8080}
-
-# Database Configuration
-spring.datasource.url=${SPRING_DATASOURCE_URL:jdbc:postgresql://localhost:5430/solontax_g1}
-spring.datasource.username=${SPRING_DATASOURCE_USERNAME:postgres}
-spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:postgres}
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-# JPA Configuration
-spring.jpa.hibernate.ddl-auto=${SPRING_JPA_DDL_AUTO:update}
-spring.jpa.show-sql=${SPRING_JPA_SHOW_SQL:true}
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.properties.hibernate.format_sql=true
-
-# Kafka Configuration
-spring.kafka.bootstrap-servers=solontax-g1-kafka:9092
-spring.kafka.consumer.group-id=solontax-group
-spring.kafka.consumer.auto-offset-reset=earliest
-spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.admin.auto-create=true
-spring.kafka.streams.application-id=solontax-streams-app
-spring.kafka.streams.replication-factor=1
-spring.kafka.streams.properties.default.key.serde=org.apache.kafka.common.serialization.Serdes$StringSerde
-spring.kafka.streams.properties.default.value.serde=org.apache.kafka.common.serialization.Serdes$StringSerde
-
-# Timezone Configuration
-spring.jpa.properties.hibernate.jdbc.time_zone=UTC
